+++ api/src/main/java/com/example/repository/DoctorScheduleRepository.java
@@ -9,8 +9,8 @@ import java.util.Optional;
 /**
  * Created by japnica on 12/25/2016.
  */
-public interface DoctorScheduleRepository extends JpaRepository<DoctorSchedule, Long> {
+public interface DoctorScheduleRepository extends JpaRepository<DoctorSchedule, Integer> {
     List<DoctorSchedule> findAll();
-    Optional<DoctorSchedule> findOneById(Long id);
+    Optional<DoctorSchedule> findOneById(Integer id);
     void delete(DoctorSchedule d);
 }
