+++ api/src/main/java/com/example/repository/ParticularBillingRepository.java
@@ -9,8 +9,8 @@ import java.util.Optional;
 /**
  * Created by japnica on 12/25/2016.
  */
-public interface ParticularBillingRepository extends JpaRepository<ParticularBilling, Long>{
+public interface ParticularBillingRepository extends JpaRepository<ParticularBilling, Integer>{
     List<ParticularBilling> findAll();
-    Optional<ParticularBilling> findOneById(Long id);
+    Optional<ParticularBilling> findOneById(Integer id);
     void delete(ParticularBilling p);
 }
