+++ api/src/main/java/com/example/repository/StaffRepository.java
@@ -10,8 +10,8 @@ import java.util.Optional;
 /**
  * Created by BiSAl MhRzn on 12/25/2016.
  */
-public interface StaffRepository extends JpaRepository<Staff,Long>{
+public interface StaffRepository extends JpaRepository<Staff,Integer>{
     List<Staff> findAll();
-    Optional<Staff> findOneById(long id);
+    Optional<Staff> findOneById(Integer id);
     void delete(Staff d);
 }
